Adding the HostName field to the Conn struct (#18)
Changing the internal-only match interface to return any parsed
hostnames.

It can be useful for implementers of Target to be able to inspect
the already-parsed SNI header (in the case of TLS) or host header
(in the case of http) to know what host was asked for by the client
in order to make additional routing decisions.  This can be used
by transparent reverse proxies where the destination is not known
in advance.